# openapi.json
{"openapi":"3.0.2","info":{"title":"Hotels booking","version":"1"},"servers":[{"url":"/v1"}],"paths":{"/auth/register":{"post":{"tags":["Auth"],"summary":"Register User","operationId":"register_user_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SUserAuth"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login User","operationId":"login_user_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SUserAuth"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/logout":{"post":{"tags":["Auth"],"summary":"Logout User","operationId":"logout_user_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/current":{"get":{"tags":["Auth"],"summary":"Read Current User","operationId":"read_current_user_auth_current_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/all-users":{"get":{"tags":["Auth"],"summary":"Read Current User","operationId":"read_current_user_auth_all_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/bookings":{"get":{"tags":["Booking"],"summary":"Get Bookings","operationId":"get_bookings_bookings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Bookings Bookings Get","type":"array","items":{"$ref":"#/components/schemas/SBookings"}}}}}}},"post":{"tags":["Booking"],"summary":"Add Booking","operationId":"add_booking_bookings_post","parameters":[{"required":true,"schema":{"title":"Room Id","type":"integer"},"name":"room_id","in":"query"},{"description":"Date format is: 2023-05-19","required":true,"schema":{"title":"Date From","type":"string","description":"Date format is: 2023-05-19","format":"date"},"name":"date_from","in":"query"},{"description":"Date format is: 2023-05-19","required":true,"schema":{"title":"Date To","type":"string","description":"Date format is: 2023-05-19","format":"date"},"name":"date_to","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/bookings/{booking_id}":{"delete":{"tags":["Booking"],"summary":"Delete Booking","operationId":"delete_booking_bookings__booking_id__delete","parameters":[{"required":true,"schema":{"title":"Booking Id","type":"integer"},"name":"booking_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/hotels/all":{"get":{"tags":["Hotels"],"summary":"Get All Hotels","operationId":"get_all_hotels_hotels_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All Hotels Hotels All Get","type":"array","items":{"$ref":"#/components/schemas/SHotels"}}}}}}}},"/hotels/id":{"get":{"tags":["Hotels"],"summary":"Get Hotel By Id","operationId":"get_hotel_by_id_hotels_id_get","parameters":[{"required":true,"schema":{"title":"Hotel Id","type":"string"},"name":"hotel_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SHotels"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/hotels/{location}":{"get":{"tags":["Hotels"],"summary":"Get Hotels By Location And Date","operationId":"get_hotels_by_location_and_date_hotels__location__get","parameters":[{"required":true,"schema":{"title":"Location","type":"string"},"name":"location","in":"path"},{"description":"Date format is: 2023-05-19","required":true,"schema":{"title":"Date From","type":"string","description":"Date format is: 2023-05-19","format":"date"},"name":"date_from","in":"query"},{"description":"Date format is: 2023-05-19","required":true,"schema":{"title":"Date To","type":"string","description":"Date format is: 2023-05-19","format":"date"},"name":"date_to","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/hotels/{hotel_id}/rooms":{"get":{"tags":["Rooms"],"summary":"Get All Rooms By Hotel Id","operationId":"get_all_rooms_by_hotel_id_hotels__hotel_id__rooms_get","parameters":[{"required":true,"schema":{"title":"Hotel Id","type":"string"},"name":"hotel_id","in":"path"},{"description":"Date format is: 2023-05-19","required":true,"schema":{"title":"Date From","type":"string","description":"Date format is: 2023-05-19","format":"date"},"name":"date_from","in":"query"},{"description":"Date format is: 2023-05-19","required":true,"schema":{"title":"Date To","type":"string","description":"Date format is: 2023-05-19","format":"date"},"name":"date_to","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All Rooms By Hotel Id Hotels  Hotel Id  Rooms Get","type":"array","items":{"$ref":"#/components/schemas/SRoomsRoomLeft"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/images/hotels":{"post":{"tags":["Images loading"],"summary":"Add Hotel Image","operationId":"add_hotel_image_images_hotels_post","parameters":[{"required":true,"schema":{"title":"Name","type":"string"},"name":"name","in":"query"}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_add_hotel_image_images_hotels_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pages/hotels":{"get":{"tags":["FrontEnd"],"summary":"Get Hotels Page","operationId":"get_hotels_page_pages_hotels_get","parameters":[{"required":true,"schema":{"title":"Location","type":"string"},"name":"location","in":"query"},{"description":"Date format is: 2023-05-19","required":true,"schema":{"title":"Date From","type":"string","description":"Date format is: 2023-05-19","format":"date"},"name":"date_from","in":"query"},{"description":"Date format is: 2023-05-19","required":true,"schema":{"title":"Date To","type":"string","description":"Date format is: 2023-05-19","format":"date"},"name":"date_to","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/import/hotels":{"post":{"tags":["Import hotels and rooms"],"summary":"Import Hotels To Database","operationId":"import_hotels_to_database_import_hotels_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/import/rooms":{"post":{"tags":["Import hotels and rooms"],"summary":"Import Rooms To Database","operationId":"import_rooms_to_database_import_rooms_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_add_hotel_image_images_hotels_post":{"title":"Body_add_hotel_image_images_hotels_post","required":["file"],"type":"object","properties":{"file":{"title":"File","type":"string","format":"binary"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"SBookings":{"title":"SBookings","required":["id","room_id","user_id","date_from","date_to","price","total_cost","total_days"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"room_id":{"title":"Room Id","type":"integer"},"user_id":{"title":"User Id","type":"integer"},"date_from":{"title":"Date From","type":"string","format":"date"},"date_to":{"title":"Date To","type":"string","format":"date"},"price":{"title":"Price","type":"integer"},"total_cost":{"title":"Total Cost","type":"integer"},"total_days":{"title":"Total Days","type":"integer"}}},"SHotels":{"title":"SHotels","required":["id","name","location","services","room_quantity","image_id"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"location":{"title":"Location","type":"string"},"services":{"title":"Services","type":"array","items":{"type":"string"}},"room_quantity":{"title":"Room Quantity","type":"integer"},"image_id":{"title":"Image Id","type":"integer"}}},"SRoomsRoomLeft":{"title":"SRoomsRoomLeft","required":["id","hotel_id","name","price","services","quantity","image_id","rooms_left"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"hotel_id":{"title":"Hotel Id","type":"integer"},"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string"},"price":{"title":"Price","type":"integer"},"services":{"title":"Services","type":"array","items":{"type":"string"}},"quantity":{"title":"Quantity","type":"integer"},"image_id":{"title":"Image Id","type":"integer"},"rooms_left":{"title":"Rooms Left","type":"integer"}}},"SUserAuth":{"title":"SUserAuth","required":["email","password"],"type":"object","properties":{"email":{"title":"Email","type":"string","format":"email"},"password":{"title":"Password","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}